refactor: archive provider policy matrices with failure postmortem
Move retired policy-layer code (provider operation/intent allowlists, matrix
combination validator, and env-var parsers) to archive/provider_policy.py with
comprehensive README documenting the observed failure (wake 25789, 2026-08-24),
root-cause analysis, and revival requirements.

- Created archive/__init__.py and archive/provider_policy.py with moved code
- Created archive/README.md with failure narrative and env var documentation
- Removed DEFAULT_ENABLED_OPERATIONS_BY_PROVIDER, DEFAULT_ENABLED_INTENTS,
  DEFAULT_ENABLED_INTENTS_BY_PROVIDER, and parser functions from server.py
- Removed ALLOWED_COMBINATIONS from models.py
- Updated test imports to reference archive module for archived-behavior tests
- All live code (_provider(), _enabled_operations(), FEATURE_POLICY) remains
