+++ src/postgres_mcp/outbound_gateway/server.py
@@ -86,15 +86,6 @@ DEFAULT_PROPERTY_ALIASES = {
     "16 north main street 16": "building:16-n-main",
 }
 DEFAULT_ENABLED_OPERATIONS = frozenset({Operation.EMAIL_SEND})
-DEFAULT_ENABLED_OPERATIONS_BY_PROVIDER = {
-    "hotpads": frozenset({Operation.EMAIL_SEND.value}),
-    "zillow": frozenset({Operation.EMAIL_SEND.value}),
-}
-DEFAULT_ENABLED_INTENTS = frozenset({IntentKind.INQUIRY_REPLY.value, IntentKind.SHOWING_OFFER.value})
-DEFAULT_ENABLED_INTENTS_BY_PROVIDER = {
-    "hotpads": frozenset({IntentKind.INQUIRY_REPLY.value, IntentKind.SHOWING_OFFER.value}),
-    "zillow": frozenset({IntentKind.INQUIRY_REPLY.value, IntentKind.SHOWING_OFFER.value}),
-}
 
 
 @dataclass(frozen=True)
@@ -274,65 +265,6 @@ def _enabled_operations() -> frozenset[Operation]:
         raise ValueError("OUTBOUND_ENABLED_OPERATIONS_JSON contains an unsupported operation") from exc
 
 
-def _enabled_operations_by_provider() -> dict[str, frozenset[str]]:
-    raw = os.environ.get("OUTBOUND_PROVIDER_OPERATIONS_JSON")
-    if raw is None:
-        return DEFAULT_ENABLED_OPERATIONS_BY_PROVIDER
-    value = json.loads(raw)
-    if not isinstance(value, dict) or not value:
-        raise ValueError("OUTBOUND_PROVIDER_OPERATIONS_JSON must be a non-empty JSON object")
-    parsed: dict[str, frozenset[str]] = {}
-    for provider, operations in value.items():
-        if (
-            not isinstance(provider, str)
-            or not provider.strip()
-            or not isinstance(operations, list)
-            or not operations
-            or not all(isinstance(item, str) for item in operations)
-        ):
-            raise ValueError("OUTBOUND_PROVIDER_OPERATIONS_JSON values must be non-empty string arrays")
-        try:
-            parsed[provider.casefold()] = frozenset(Operation(item).value for item in operations)
-        except ValueError as exc:
-            raise ValueError("OUTBOUND_PROVIDER_OPERATIONS_JSON contains an unsupported operation") from exc
-    return parsed
-
-
-def _enabled_intents() -> frozenset[str]:
-    raw = os.environ.get("OUTBOUND_ENABLED_INTENTS_JSON")
-    if raw is None:
-        return DEFAULT_ENABLED_INTENTS
-    value = json.loads(raw)
-    if not isinstance(value, list) or not value or not all(isinstance(item, str) for item in value):
-        raise ValueError("OUTBOUND_ENABLED_INTENTS_JSON must be a non-empty string array")
-    try:
-        return frozenset(IntentKind(item).value for item in value)
-    except ValueError as exc:
-        raise ValueError("OUTBOUND_ENABLED_INTENTS_JSON contains an unsupported intent") from exc
-
-
-def _enabled_intents_by_provider() -> dict[str, frozenset[str]]:
-    raw = os.environ.get("OUTBOUND_PROVIDER_INTENTS_JSON")
-    if raw is None:
-        return DEFAULT_ENABLED_INTENTS_BY_PROVIDER
-    value = json.loads(raw)
-    if not isinstance(value, dict) or not value:
-        raise ValueError("OUTBOUND_PROVIDER_INTENTS_JSON must be a non-empty JSON object")
-    parsed: dict[str, frozenset[str]] = {}
-    for provider, intents in value.items():
-        if (
-            not isinstance(provider, str)
-            or not provider.strip()
-            or not isinstance(intents, list)
-            or not intents
-            or not all(isinstance(item, str) for item in intents)
-        ):
-            raise ValueError("OUTBOUND_PROVIDER_INTENTS_JSON values must be non-empty string arrays")
-        try:
-            parsed[provider.casefold()] = frozenset(IntentKind(item).value for item in intents)
-        except ValueError as exc:
-            raise ValueError("OUTBOUND_PROVIDER_INTENTS_JSON contains an unsupported intent") from exc
-    return parsed
 
 
 def _bearer_headers(name: str) -> dict[str, str]:
@@ -538,11 +470,6 @@ async def build_runtime() -> GatewayRuntime:
             DEFAULT_PROPERTY_ALIASES,
         ),
         conversation_aliases=_json_mapping("OUTBOUND_CONVERSATION_ALIASES_JSON", {}),
-        # _enabled_operations_by_provider() / _enabled_intents() /
-        # _enabled_intents_by_provider() are no longer wired into
-        # RoutingPolicy -- Task 4 removed the gates they fed; the parsers
-        # and DEFAULT_ENABLED_* constants stay for now (still exercised by
-        # test_server.py) and are archived in Task 5.
     )
     context_repository = OutboundGatewayRepository(driver)
     store = PostgresActionStore(driver)
