+++ tests/unit/outbound_gateway/test_server.py
@@ -16,17 +16,17 @@ from postgres_mcp.outbound_gateway.models import ActionRole
 from postgres_mcp.outbound_gateway.models import Operation
 from postgres_mcp.outbound_gateway.models import PublicResult
 from postgres_mcp.outbound_gateway.models import PublicStatus
+from postgres_mcp.outbound_gateway.archive.provider_policy import DEFAULT_ENABLED_INTENTS
+from postgres_mcp.outbound_gateway.archive.provider_policy import DEFAULT_ENABLED_INTENTS_BY_PROVIDER
+from postgres_mcp.outbound_gateway.archive.provider_policy import DEFAULT_ENABLED_OPERATIONS_BY_PROVIDER
+from postgres_mcp.outbound_gateway.archive.provider_policy import _enabled_intents_by_provider
+from postgres_mcp.outbound_gateway.archive.provider_policy import _enabled_operations_by_provider
 from postgres_mcp.outbound_gateway.server import DEFAULT_EMAIL_CC_BY_SOURCE
 from postgres_mcp.outbound_gateway.server import DEFAULT_EMAIL_SENDER_DOMAINS
-from postgres_mcp.outbound_gateway.server import DEFAULT_ENABLED_INTENTS
-from postgres_mcp.outbound_gateway.server import DEFAULT_ENABLED_INTENTS_BY_PROVIDER
-from postgres_mcp.outbound_gateway.server import DEFAULT_ENABLED_OPERATIONS_BY_PROVIDER
 from postgres_mcp.outbound_gateway.server import DEFAULT_PROPERTY_ALIASES
 from postgres_mcp.outbound_gateway.server import TENANTCLOUD_ORIGIN
 from postgres_mcp.outbound_gateway.server import FeaturePolicy
 from postgres_mcp.outbound_gateway.server import _bearer_headers
-from postgres_mcp.outbound_gateway.server import _enabled_intents_by_provider
-from postgres_mcp.outbound_gateway.server import _enabled_operations_by_provider
 from postgres_mcp.outbound_gateway.server import _reject_tenantcloud_origin_overrides
 from postgres_mcp.outbound_gateway.server import _tenantcloud_adapters
 from postgres_mcp.outbound_gateway.server import _tenantcloud_enabled
